requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
telegraph==2.2.0
aiohttp==3.9.1